sprint 3

+++ src/main/java/co/edu/udea/LupinosDEV/repositories/MovimientoDineroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovimientoDineroRepository extends JpaRepository<MovimientoDinero,Integer>{
+public interface MovimientoDineroRepository extends JpaRepository<MovimientoDinero,Long>{
 }
